requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
openai
langchain-community
langchain-core
langchain-text-splitters
requests
beautifulsoup4
pydantic
faiss-cpu
jinja2
python-multipart
pypdf
llama-index
snowflake-connector-python